requirements.txt
python-pptx>=0.6.21
markdown>=3.4.0
markdown-it-py>=3.0.0
